+++ lib/maze/api/appium/app_manager.rb
@@ -17,15 +17,18 @@ module Maze
 
           @driver.activate_app(@driver.app_id)
           true
-        rescue Selenium::WebDriver::Error::ServerError => e
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to activate app: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
 
-        # Terminates the app
+        # Terminates the app.  If terminate fails then clients may wish to ty the legacy close method, so an option
+        # is provided to not fail the Appium driver.
+        # @param fail_driver [Boolean] Whether to fail the Appium driver if the app cannot be terminated
         # @returns [Boolean] Whether the app was successfully closed
-        def terminate
+        def terminate(fail_driver = true)
           if failed_driver?
             $logger.error 'Cannot terminate the app - Appium driver failed.'
             return false
@@ -33,9 +36,28 @@ module Maze
 
           @driver.terminate_app(@driver.app_id)
           true
-        rescue Selenium::WebDriver::Error::ServerError => e
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to terminate app: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message) if fail_driver
+          raise e
+        end
+
+        # Instructs Appium to background the app
+        # @param seconds [Integers] The number of seconds to background the app for, or -1 for indefinitely
+        # @returns [Boolean] Whether the instruction to Appium was successfully made
+        def background(seconds = -1)
+          if failed_driver?
+            $logger.error 'Cannot background the app - Appium driver failed.'
+            return false
+          end
+
+          @driver.background_app(seconds)
+          true
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to background app: #{e.message}"
+          # Assume the remote appium session has stopped, so crash out of the session
+          fail_driver(e.message)
           raise e
         end
 
@@ -49,9 +71,10 @@ module Maze
 
           @driver.launch_app
           true
-        rescue Selenium::WebDriver::Error::ServerError => e
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to launch app: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
 
@@ -65,9 +88,10 @@ module Maze
 
           @driver.close_app
           true
-        rescue Selenium::WebDriver::Error::ServerError => e
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to close app: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
 
@@ -80,9 +104,10 @@ module Maze
           end
 
           @driver.app_state(@driver.app_id)
-        rescue Selenium::WebDriver::Error::ServerError => e
+        rescue Selenium::WebDriver::Error::ServerError, Selenium::WebDriver::Error::UnknownError => e
+          $logger.error "Failed to get app state: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
       end