+++ lib/maze/api/appium/device_manager.rb
@@ -19,8 +19,44 @@ module Maze
           @driver.unlock
           true
         rescue Selenium::WebDriver::Error::ServerError => e
+          $logger.error "Failed to unlock the device: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
+          raise e
+        end
+
+        # Presses the Back button.
+        # @returns [Boolean] Success status
+        def back
+          if failed_driver?
+            $logger.error 'Cannot press the Back button - Appium driver failed.'
+            return false
+          end
+
+          @driver.back
+          true
+        rescue Selenium::WebDriver::Error::ServerError => e
+          $logger.error "Failed to press back: #{e.message}"
+          # Assume the remote appium session has stopped, so crash out of the session
+          fail_driver(e.message)
+          raise e
+        end
+
+        # Gets logs from the device.
+        # @param log_type [String] The type pf log to get as recognised by Appium, such as 'syslog'
+        #
+        # @returns [Array, nil] Array of Selenium::WebDriver::LogEntry, or nil if the driver has failed
+        def get_log(log_type)
+          if failed_driver?
+            $logger.error 'Cannot get logs - Appium driver failed.'
+            return nil
+          end
+
+          @driver.get_log(log_type)
+        rescue Selenium::WebDriver::Error::ServerError => e
+          $logger.error "Failed to get logs: #{e.message}"
+          # Assume the remote appium session has stopped, so crash out of the session
+          fail_driver(e.message)
           raise e
         end
 
@@ -36,8 +72,9 @@ module Maze
           @driver.set_rotation(orientation)
           true
         rescue Selenium::WebDriver::Error::ServerError => e
+          $logger.error "Failed to set the device rotation: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
 
@@ -51,8 +88,9 @@ module Maze
 
           JSON.generate(@driver.device_info)
         rescue Selenium::WebDriver::Error::ServerError => e
+          $logger.error "Failed to get the device info: #{e.message}"
           # Assume the remote appium session has stopped, so crash out of the session
-          fail_driver
+          fail_driver(e.message)
           raise e
         end
       end