+++ lib/maze/hooks/appium_hooks.rb
@@ -4,12 +4,6 @@ module Maze
     # Hooks for Appium mode use
     class AppiumHooks < InternalHooks
 
-      APPIUM_DRIVER_FAILED_ERRORS = [
-        Selenium::WebDriver::Error::UnknownError,
-        Selenium::WebDriver::Error::ServerError,
-        Selenium::WebDriver::Error::WebDriverError
-      ]
-
       @client
 
       def before_all
@@ -30,29 +24,30 @@ module Maze
       end
 
       def after(scenario)
-        if scenario.failed?
-          Maze.driver = nil if driver_has_failed?
-          $logger.warn('The appium driver has failed, removing it to prevent future errors')
-        end
+        $logger.debug "Appium after hook"
 
+        manager = Maze::Api::Appium::AppManager.new
         if Maze.config.os == 'macos'
-          # Close the app - without the sleep, launching the app for the next scenario intermittently fails
+          # Close the app - without the sleep launching the app for the next scenario intermittently fails
           system("killall -KILL #{Maze.config.app} && sleep 1")
         elsif [:bb, :bs, :local].include? Maze.config.farm
-          # Reset the server to ensure that test fixtures cannot fetch
-          # commands from the previous scenario (in idempotent mode).
+          close_fallback = Maze.config.appium_version && Maze.config.appium_version.to_f < 2.0
           begin
-            Maze.driver.terminate_app Maze.driver.app_id
-          rescue Selenium::WebDriver::Error::UnknownError, Selenium::WebDriver::Error::InvalidSessionIdError
-            if Maze.config.appium_version && Maze.config.appium_version.to_f < 2.0
+            manager.terminate(!close_fallback)
+          rescue Selenium::WebDriver::Error::UnknownError, Selenium::WebDriver::Error::InvalidSessionIdError => e
+            if close_fallback
+              Bugsnag.notify(e)
               $logger.warn 'terminate_app failed, using the slower but more forceful close_app instead'
-              Maze.driver.close_app
+              manager.close
             else
-              $logger.warn 'terminate_app failed, future errors may occur if the application did not close remotely'
+              $logger.warn 'terminate_app failed, future errors may occur if the application did not close properly'
             end
           end
+
+          # Reset the server before relaunching the app to ensure that test fixtures cannot fetch
+          # commands from the previous scenario (in idempotent mode).
           Maze::Server.reset!
-          Maze.driver.activate_app Maze.driver.app_id
+          manager.activate
         end
       rescue => error
         # Notify and re-raise for Cucumber to handle
@@ -77,30 +72,6 @@ module Maze
           @client.log_run_outro
           $logger.info 'Stopping the Appium session'
           @client.stop_session
-        end
-      end
-
-      private
-
-      def driver_has_failed?
-        return false if Maze::ErrorCaptor.empty?
-
-        Maze::ErrorCaptor.classes.any? do |error_class|
-          APPIUM_DRIVER_FAILED_ERRORS.include?(error_class)
-        end
-      end
-
-      # Pulls the device logs using Appium and writes them to file in the maze_output folder
-      def write_device_logs(scenario)
-        log_name = case Maze::Helper.get_current_platform
-                   when 'android'
-                     'logcat'
-                   when 'ios'
-                     'syslog'
-                   end
-        unless Maze.driver.nil?
-          logs = Maze.driver.get_log(log_name)
-          Maze::MazeOutput.new(scenario).write_device_logs(logs)
         end
       end
     end