+++ test/e2e/appium-api/features/steps/steps.rb
@@ -22,6 +22,10 @@ When('I unlock the device') do
   Maze::Api::Appium::DeviceManager.new.unlock
 end
 
+When('I press the back button') do
+  Maze::Api::Appium::DeviceManager.new.back
+end
+
 When('I set the device rotation to {string}') do |orientation|
   Maze::Api::Appium::DeviceManager.new.set_rotation(orientation.to_sym)
 end
@@ -30,3 +34,14 @@ When('I log the device info') do
   info = Maze::Api::Appium::DeviceManager.new.info
   $logger.info "Device info is: #{info}"
 end
+
+When('I get the device logs') do
+  log_type = if Maze.config.device.include?('ANDROID')
+               'logcat'
+             else
+               'syslog'
+             end
+  logs = Maze::Api::Appium::DeviceManager.new.get_log(log_type)
+  $logger.info "#{logs.length} device logs entries returned"
+end
+