+++ test/unit/maze/api/appium/app_manager_test.rb
@@ -51,13 +51,21 @@ module Maze
           assert_false(@manager.activate)
         end
 
-        def test_state_server_error
+        def test_background_failed_driver
+          @mock_driver.expects(:failed?).returns(true)
+          $logger.expects(:error).with("Cannot background the app - Appium driver failed.")
+
+          assert_false(@manager.background)
+        end
+
+        def test_state_unknown_error
           @mock_driver.expects(:failed?).returns(false)
           @mock_driver.expects(:app_id).returns('app1')
           @mock_driver.expects(:fail_driver)
-          @mock_driver.expects(:app_state).with('app1').raises(Selenium::WebDriver::Error::ServerError, 'Timeout')
+          @mock_driver.expects(:app_state).with('app1').raises(Selenium::WebDriver::Error::UnknownError, 'Timeout')
+          $logger.expects(:error).with("Failed to get app state: Timeout")
 
-          error = assert_raises Selenium::WebDriver::Error::ServerError do
+          error = assert_raises Selenium::WebDriver::Error::UnknownError do
             @manager.state
           end
           assert_equal 'Timeout', error.message
@@ -67,6 +75,7 @@ module Maze
           @mock_driver.expects(:failed?).returns(false)
           @mock_driver.expects(:fail_driver)
           @mock_driver.expects(:close_app).raises(Selenium::WebDriver::Error::ServerError, 'Timeout')
+          $logger.expects(:error).with("Failed to close app: Timeout")
 
           error = assert_raises Selenium::WebDriver::Error::ServerError do
             @manager.close
@@ -78,9 +87,22 @@ module Maze
           @mock_driver.expects(:failed?).returns(false)
           @mock_driver.expects(:fail_driver)
           @mock_driver.expects(:launch_app).raises(Selenium::WebDriver::Error::ServerError, 'Timeout')
+          $logger.expects(:error).with("Failed to launch app: Timeout")
 
           error = assert_raises Selenium::WebDriver::Error::ServerError do
             @manager.launch
+          end
+          assert_equal 'Timeout', error.message
+        end
+
+        def test_background_server_error
+          @mock_driver.expects(:failed?).returns(false)
+          @mock_driver.expects(:fail_driver)
+          @mock_driver.expects(:background_app).with(-1).raises(Selenium::WebDriver::Error::ServerError, 'Timeout')
+          $logger.expects(:error).with("Failed to background app: Timeout")
+
+          error = assert_raises Selenium::WebDriver::Error::ServerError do
+            @manager.background
           end
           assert_equal 'Timeout', error.message
         end